Add tests for custom function template
This test splits off the cpp tests into 2 copies. The first set is
left as is (using normal function pointers): fff_test_cpp &
fff_test_glob_cpp. While the other (fff_test_cpp_custom_fn &
fff_test_glob_cpp_custom_fn) uses std::function for the custom_fn
signature.

Some restructuring was needed for this. The main issue is that
global_fakes.c was used to define the declared global fakes which
needed to use std::function (C++). The DEFINE_FAKE_* calls were moved
to a global_fakes.include and the global fakes was split off to
global_fakes.c and global_fakes_cpp.cpp.

Next custom_function.hpp was auto-included to the compilation. To do
that, the Makefile targets which match %_custom_fn.o were mapped to
%.cpp and the extra header included as a g++ -include flag. This
avoided creating extra files
