+++ test/global_fakes_cpp.cpp
@@ -1,0 +1,4 @@
+#include "global_fakes.hpp"
+#include <cstring> // for memcpy
+
+#include "global_fakes.include"
