+++ test/test_cases.include
@@ -244,9 +244,11 @@ void voidfunc1outparam_custom_fake3(char *a)
 
 TEST_F(FFFTestSuite, custom_fake_sequence_not_exausthed)
 {
-  void (*custom_fakes[])(char *) = {voidfunc1outparam_custom_fake1,
-                                    voidfunc1outparam_custom_fake2,
-                                    voidfunc1outparam_custom_fake3};
+  CUSTOM_FFF_FUNCTION_TEMPLATE(void, custom_fakes[], char *) = {
+      voidfunc1outparam_custom_fake1,
+      voidfunc1outparam_custom_fake2,
+      voidfunc1outparam_custom_fake3
+  };
   char a = 'a';
 
   SET_CUSTOM_FAKE_SEQ(voidfunc1outparam, custom_fakes, 3);
@@ -304,9 +306,10 @@ long custom_longfunc3(void)
 
 TEST_F(FFFTestSuite, custom_fake_seq_return_values_saved_in_history)
 {
-  long (*custom_fakes[])(void) = {custom_longfunc1,
-                                  custom_longfunc2,
-                                  custom_longfunc3};
+  CUSTOM_FFF_FUNCTION_TEMPLATE(long, custom_fakes[], void) = {
+			custom_longfunc1,
+			custom_longfunc2,
+			custom_longfunc3};
 
   SET_CUSTOM_FAKE_SEQ(longfunc0, custom_fakes, 3);
 
@@ -321,9 +324,10 @@ TEST_F(FFFTestSuite, custom_fake_seq_return_values_saved_in_history)
 
 TEST_F(FFFTestSuite, custom_fake_sequence_exhausted)
 {
-  void (*custom_fakes[])(char *) = {voidfunc1outparam_custom_fake1,
-                                    voidfunc1outparam_custom_fake2,
-                                    voidfunc1outparam_custom_fake3};
+  CUSTOM_FFF_FUNCTION_TEMPLATE(void, custom_fakes[], char *) = {
+			voidfunc1outparam_custom_fake1,
+			voidfunc1outparam_custom_fake2,
+			voidfunc1outparam_custom_fake3};
   char a = 'a';
 
   SET_CUSTOM_FAKE_SEQ(voidfunc1outparam, custom_fakes, 3);
@@ -399,9 +403,10 @@ int valuefunc3var_custom_fake3(const char *str, int a, va_list vl)
 
 TEST_F(FFFTestSuite, vararg_custom_fake_sequence_not_exhausted)
 {
-	int (*custom_fakes[])(const char *, int, va_list) = {valuefunc3var_custom_fake1, 
-                                                         valuefunc3var_custom_fake2,
-	                                                     valuefunc3var_custom_fake3};
+	CUSTOM_FFF_FUNCTION_TEMPLATE(int, custom_fakes[], const char *, int,
+			va_list) = {valuefunc3var_custom_fake1,
+									valuefunc3var_custom_fake2,
+									valuefunc3var_custom_fake3};
 	SET_CUSTOM_FAKE_SEQ(valuefunc3var, custom_fakes, 3);
 	int a = 1;
 	ASSERT_EQ(valuefunc3var("a", a, 2, 3, 4, 0), 10);
@@ -411,9 +416,10 @@ TEST_F(FFFTestSuite, vararg_custom_fake_sequence_not_exhausted)
 
 TEST_F(FFFTestSuite, vararg_custom_fake_seq_return_values_saved_in_history)
 {
-	int (*custom_fakes[])(const char *, int, va_list) = {valuefunc3var_custom_fake1,
-                                                         valuefunc3var_custom_fake2,
-	                                                     valuefunc3var_custom_fake3};
+	CUSTOM_FFF_FUNCTION_TEMPLATE(int, custom_fakes[], const char *, int,
+                                     va_list) = {valuefunc3var_custom_fake1,
+                                           valuefunc3var_custom_fake2,
+                                           valuefunc3var_custom_fake3};
 	SET_CUSTOM_FAKE_SEQ(valuefunc3var, custom_fakes, 3);
 	int a = 1;
 	valuefunc3var("a", a, 2, 3, 4, 0);
@@ -426,9 +432,10 @@ TEST_F(FFFTestSuite, vararg_custom_fake_seq_return_values_saved_in_history)
 
 TEST_F(FFFTestSuite, vararg_custom_fake_sequence_exhausted)
 {
-	int (*custom_fakes[])(const char *, int, va_list) = {valuefunc3var_custom_fake1,
-                                                         valuefunc3var_custom_fake2,
-	                                                     valuefunc3var_custom_fake3};
+	CUSTOM_FFF_FUNCTION_TEMPLATE(int, custom_fakes[], const char *, int,
+                               va_list) = {valuefunc3var_custom_fake1,
+                                           valuefunc3var_custom_fake2,
+                                           valuefunc3var_custom_fake3};
 	SET_CUSTOM_FAKE_SEQ(valuefunc3var, custom_fakes, 3);
 	int a = 1;
 	ASSERT_EQ(valuefunc3var("a", a, 2, 3, 4, 0), 10);
@@ -440,9 +447,10 @@ TEST_F(FFFTestSuite, vararg_custom_fake_sequence_exhausted)
 
 TEST_F(FFFTestSuite, vararg_custom_fake_sequence_reset)
 {
-	int (*custom_fakes[])(const char *, int, va_list) = {valuefunc3var_custom_fake1,
-                                                         valuefunc3var_custom_fake2,
-	                                                     valuefunc3var_custom_fake3};
+	CUSTOM_FFF_FUNCTION_TEMPLATE(int, custom_fakes[], const char *, int,
+                               va_list) = {valuefunc3var_custom_fake1,
+                                           valuefunc3var_custom_fake2,
+                                           valuefunc3var_custom_fake3};
 	SET_CUSTOM_FAKE_SEQ(valuefunc3var, custom_fakes, 3);
 	int a = 1;
 	ASSERT_EQ(valuefunc3var("a", a, 2, 3, 4, 0), 10);
